Refactor locations from initializer to Bibdata, cache in Rails.cache. refs #656
to allow stubbing bibdata/locations api call with webmock
