--- config/initializers/locations_initializer.rb
@@ -1,13 +1,0 @@
-require 'holding_locations'
-
-module Orangelight
-  def locations
-    @locations = HoldingLocations.load if @locations.blank?
-    @locations
-  end
-
-  module_function :locations
-end
-
-# initialize holding locations data
-Orangelight.locations
